Rename zero_copy -> zerocopy.
